requirements.txt
selenium==4.1.0
beautifulsoup4==4.9.3
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
